+++ coachmark/src/main/java/com/mercadolibre/android/andesui/coachmark/model/AndesWalkthroughCoachmark.kt
@@ -1,5 +1,5 @@
 package com.mercadolibre.android.andesui.coachmark.model
-import android.support.v4.widget.NestedScrollView
+import androidx.core.widget.NestedScrollView
 import android.view.View
 import java.io.Serializable