+++ coachmark/src/main/java/com/mercadolibre/android/andesui/coachmark/view/CoachmarkOverlay.kt
@@ -6,7 +6,7 @@ import android.graphics.PorterDuff
 import android.graphics.PorterDuffXfermode
 import android.graphics.RectF
 import android.graphics.Canvas
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 import android.util.AttributeSet
 import android.view.View
 import com.mercadolibre.android.andesui.coachmark.R