+++ components/src/main/java/com/mercadolibre/android/andesui/color/ColorKtx.kt
@@ -1,8 +1,8 @@
 package com.mercadolibre.android.andesui.color
 
 import android.content.Context
-import android.support.annotation.ColorInt
-import android.support.v4.content.ContextCompat
+import androidx.annotation.ColorInt
+import androidx.core.content.ContextCompat
 
 @ColorInt
 internal fun Int.toColor(context: Context): Int {