+++ components/src/test/java/com/mercadolibre/android/andesui/badge/AndesBadgeDotConfigurationTest.kt
@@ -16,7 +16,7 @@ import org.robolectric.RuntimeEnvironment
 import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
-@Config(constants = BuildConfig::class, sdk = [Build.VERSION_CODES.LOLLIPOP])
+@Config(sdk = [Build.VERSION_CODES.LOLLIPOP])
 class AndesBadgeDotConfigurationTest {
 
     private var context = RuntimeEnvironment.application