+++ components/src/test/java/com/mercadolibre/android/andesui/progress/AndesProgressSizeTest.kt
@@ -15,7 +15,7 @@ import org.robolectric.RuntimeEnvironment
 import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
-@Config(constants = BuildConfig::class, sdk = [Build.VERSION_CODES.LOLLIPOP])
+@Config(sdk = [Build.VERSION_CODES.LOLLIPOP])
 class AndesProgressSizeTest {
     private var andesXLargeProgressSize = Mockito.spy(AndesXLargeProgressSize())
     private var andesLargeProgressSize = Mockito.spy(AndesLargeProgressSize())