+++ components/src/test/java/com/mercadolibre/android/andesui/textfield/TextFieldMaskWatcherTest.kt
@@ -12,7 +12,7 @@ import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
-@Config(constants = BuildConfig::class, sdk = [Build.VERSION_CODES.LOLLIPOP])
+@Config(sdk = [Build.VERSION_CODES.LOLLIPOP])
 class TextFieldMaskWatcherTest {
 
     companion object {