+++ components/src/test/java/com/mercadolibre/android/andesui/textfieldcode/AndesTextfieldBoxWatcherTest.kt
@@ -21,7 +21,7 @@ import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
-@Config(constants = BuildConfig::class, sdk = [Build.VERSION_CODES.LOLLIPOP])
+@Config(sdk = [Build.VERSION_CODES.LOLLIPOP])
 class AndesTextfieldBoxWatcherTest {
     @Mock
     private lateinit var focusManagement: AndesCodeFocusManagement