+++ demoapp/src/main/java/com/mercadolibre/android/andesui/demoapp/TextfieldInflateHelper.kt
@@ -1,7 +1,7 @@
 package com.mercadolibre.android.andesui.demoapp
 
 import android.content.Context
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 import android.text.Editable
 import android.text.InputType
 import android.text.TextWatcher