+++ demoapp/src/main/java/com/mercadolibre/android/andesui/demoapp/feature/CardShowcaseActivity.kt
@@ -2,9 +2,9 @@ package com.mercadolibre.android.andesui.demoapp.feature
 
 import android.content.Context
 import android.os.Bundle
-import android.support.v4.view.PagerAdapter
-import android.support.v4.view.ViewPager
-import android.support.v7.app.AppCompatActivity
+import androidx.viewpager.widget.PagerAdapter
+import androidx.viewpager.widget.ViewPager
+import androidx.appcompat.app.AppCompatActivity
 import android.util.TypedValue
 import android.view.LayoutInflater
 import android.view.View