+++ demoapp/src/main/java/com/mercadolibre/android/andesui/demoapp/feature/andesbottomsheet/BottomSheetShowcaseActivity.kt
@@ -3,7 +3,7 @@ package com.mercadolibre.android.andesui.demoapp.feature.andesbottomsheet
 import android.app.Activity
 import android.os.Bundle
 import android.os.Handler
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.view.Gravity
 import android.view.View
 import android.view.ViewGroup