+++ demoapp/src/main/java/com/mercadolibre/android/andesui/demoapp/feature/utils/PageIndicator.java
@@ -2,10 +2,10 @@ package com.mercadolibre.android.andesui.demoapp.feature.utils;
 
 import android.content.Context;
 import android.graphics.PorterDuff;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.view.PagerAdapter;
-import android.support.v4.view.ViewPager;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.viewpager.widget.PagerAdapter;
+import androidx.viewpager.widget.ViewPager;
 import android.util.AttributeSet;
 import android.view.LayoutInflater;
 import android.view.View;